Minor changes
Removed the .DB file and added it to the gitignore, theese files do not need to be in the repo.
ServerHelpers are now their own header and not part of the Server class.
Session timeout handling
Html changes.
Server now runs on another thread, to not block the main thread.

+++ Config.h
@@ -20,8 +20,6 @@ X(Methog, "Hedgehog, but with Metal.") \
 X(Ratmen, "A stealth-sneak rat.")
 
 
-
-
 namespace Config {
     namespace Monsters {
         // The base chance to hit. Defense will subtract from this, attack will add.
@@ -37,7 +35,7 @@ namespace Config {
     }
 
     namespace Server {
-        inline constexpr float SessionTimeout = 60;
+        inline constexpr long SessionTimeout = 60;
     }
 }
